[Tag,MultiTag] Merge classes with mixins
